homework4/pyproject.toml
[tool.poetry]
name = "assembler"
version = "0.1.0"
description = ""
authors = ["dmitry <kalibdune@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
